renamed main application from App to OozieUI
